### What problem does this PR solve? <!--add issue link with summary if exists-->
pingcap/tidb#9635

### What is changed and how it works?
GetCharsetDesc  now get desc from inited variable descs.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test

Code changes
 - Has exported function/method change

Related changes

 - Need to update go.mod in tidb