@hhxcc 2019-03-22
Benchmark GetCharsetDesc.

Before:
```
GO111MODULE=on go test -run GetCharsetDesc   -bench Benchmark  github.com/pingcap/parser/charset
goos: darwin
goarch: amd64
pkg: github.com/pingcap/parser/charset
BenchmarkGetCharsetDesc-8       20000000                54.5 ns/op
PASS
ok      github.com/pingcap/parser/charset       1.160s
```


After:
```
GO111MODULE=on go test -run GetCharsetDesc   -bench Benchmark  github.c
om/pingcap/parser/charset
goos: darwin
goarch: amd64
pkg: github.com/pingcap/parser/charset
BenchmarkGetCharsetDesc-8       100000000               13.4 ns/op
PASS
ok      github.com/pingcap/parser/charset       1.361s
```